refactor/portfolio: moved layout components to layout folder, added react rounter - navigation to pages via navigation buttons.
