+++ portfolio-lorna/declarations.d.ts
@@ -1,1 +1,2 @@
-declare module "*.module.css";
+declare module '*.module.css';
+declare module '*.png';
