+++ portfolio-lorna/package.json
@@ -14,7 +14,8 @@
     },
     "dependencies": {
         "react": "^17.0.2",
-        "react-dom": "^17.0.2"
+        "react-dom": "^17.0.2",
+        "react-router-dom": "^6.3.0"
     },
     "devDependencies": {
         "@babel/core": "^7.17.9",
