--- portfolio-lorna/src/components/Contact/CTAButton.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const CTAButton: React.VFC = (): JSX.Element => {
-    return <div>CTAButton</div>;
-};
-
-export default CTAButton;
