+++ portfolio-lorna/src/components/Layout/Banner/Banner.tsx
