--- portfolio-lorna/src/components/Layout/landing-page/MainHeader/MainHeader.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const MainHeader = () => {
-    return <div>MainHeader</div>;
-};
-
-export default MainHeader;
