--- portfolio-lorna/src/components/Layout/landing-page/MainIntroduction/MainIntroduction.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const MainIntroduction = () => {
-    return <div>MainIntroduction</div>;
-};
-
-export default MainIntroduction;
