--- portfolio-lorna/src/components/Projects/ProjectCard.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProjectCard = () => {
-  return <div>ProjectCard</div>;
-};
-
-export default ProjectCard;
