--- portfolio-lorna/src/components/Projects/ProjectDescribtion.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProjectDescribtion = () => {
-  return <div>ProjectDescribtion</div>;
-};
-
-export default ProjectDescribtion;
