--- portfolio-lorna/src/components/Projects/ProjectIntroduction.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const ProjectIntroduction = () => {
-  return <div>ProjectIntroduction</div>;
-};
-
-export default ProjectIntroduction;
