+++ portfolio-lorna/src/components/UI/Button.tsx
@@ -1,18 +1,20 @@
-import React from "react";
+import React, { MouseEventHandler } from 'react';
 
 type Props = {
     btnClass: string;
-    children: React.ReactChild;
-    type: "submit" | "reset" | "button";
+    children: React.ReactNode;
+    type: 'submit' | 'reset' | 'button';
+    action: MouseEventHandler<HTMLButtonElement>;
 };
 
 const Button: React.FC<Props> = ({
     btnClass,
     children,
-    type = "button",
+    type = 'button',
+    action,
 }): JSX.Element => {
     return (
-        <button className={btnClass} type={type}>
+        <button className={btnClass} type={type} onClick={action}>
             {children}
         </button>
     );
