+++ portfolio-lorna/src/components/UI/Image.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import Lorna from "../../assets/images/lorna.png";
+import React from 'react';
+import Lorna from '../../assets/images/lorna.png';
 
 const Image: React.VFC = (): JSX.Element => {
     return (
