+++ portfolio-lorna/src/views/Landing/Landing.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import Banner from '../../components/Layout/Banner/Banner';
+import Contact from '../../components/Layout/Contact/Contact';
+import Portfolio from '../../components/Layout/Portfolio/Portfolio';
+
+const Landing: React.VFC = (): JSX.Element => {
+    return (
+        <>
+            <Portfolio />
+            <Banner />
+            <Contact />
+        </>
+    );
+};
+
+export default Landing;
