+++ portfolio-lorna/src/views/Projects/Projects.tsx
@@ -1,5 +1,5 @@
-import React from "react";
-import Diorama from "../components/Visuals/Diorama/Diorama";
+import React from 'react';
+import Diorama from '../../components/Visuals/Diorama/Diorama';
 
 const Projects: React.FC = (): JSX.Element => {
     return (
