+++ portfolio-lorna/src/views/Skills/Skills.tsx
@@ -1,5 +1,6 @@
-import React from "react";
-import Diorama from "../components/Visuals/Diorama/Diorama";
+import React from 'react';
+
+import Diorama from '../../components/Visuals/Diorama/Diorama';
 
 const Skills: React.FC = (): JSX.Element => {
     return (
